рабочая версия (По программе все, останутся тесты и профили)

+++ src/main/java/ru/dimas/weather/controller/LogoutController.java
@@ -4,12 +4,10 @@ import jakarta.servlet.http.HttpSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import ru.dimas.weather.service.withdb.SessionService;
-import ru.dimas.weather.service.withdb.UserService;
 
 import java.util.UUID;
 
@@ -19,11 +17,9 @@ public class LogoutController {
 
     private static final Logger logger = LoggerFactory.getLogger(WeatherController.class);
     private final SessionService sessionService;
-//    private final UserService userService;
 
-    public LogoutController(SessionService sessionService, UserService userService) {
+    public LogoutController(SessionService sessionService) {
         this.sessionService = sessionService;
-//        this.userService = userService;
     }
 
     @GetMapping
@@ -31,12 +27,16 @@ public class LogoutController {
         logger.info("show Index page");
         return "index"; // Имя шаблона
     }
+
     @PostMapping
-    public String logoutUser(HttpSession httpSession, Model model) {
+    public String logoutUser(HttpSession httpSession) {
         logger.info("logoutUser with id: {}", httpSession.getAttribute("userId"));
-        sessionService.deleteSession((UUID) httpSession.getAttribute("sessionId"));
-        httpSession.removeAttribute("userId");
-        httpSession.removeAttribute("sessionId");
+        UUID sessionId = (UUID) httpSession.getAttribute("sessionId");
+        if (sessionId != null) {
+            sessionService.deleteSession((UUID) httpSession.getAttribute("sessionId"));
+            httpSession.removeAttribute("userId");
+            httpSession.removeAttribute("sessionId");
+        }
         logger.info("user has logged out");
         return "index"; // Имя шаблона
     }
